requirements.txt
# Required packages to run the script

pillow